refactor: remove duplicate interfaces
