+++ src/hooks/usePlatforms.ts
@@ -2,7 +2,7 @@ import platforms from "../data/platforms";
 import {useQuery} from "@tanstack/react-query";
 import apiClient, { FetchResponse } from "../services/api-client";
 
-interface Platform {
+export interface Platform {
   id: number;
   name: string;
   slug: string;
